refactor: simplify library slice.
