+++ lib/features/library/providers/category_novels_family.dart
@@ -1,7 +1,7 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:nacht/common/common.dart';
 
-import '../../domain/domain.dart';
+import '../services/watch_category_novels.dart';
 
 final categoryNovelsFamily =
     StreamProvider.autoDispose.family<List<NovelData>, int>((ref, categoryId) {
