+++ lib/features/library/views/library_page.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../presentation.dart';
+import '../providers/library_view_provider.dart';
+import '../widgets/singular_library_display.dart';
+import '../widgets/tabular_library_display.dart';
 
 class LibraryPage extends HookConsumerWidget {
   const LibraryPage({Key? key}) : super(key: key);
