+++ lib/features/library/widgets/singular_library_display.dart
