+++ lib/features/library/widgets/tabular_library_display.dart
